fix #96 - Add experimental support to Immutables

+++ jpo-test/jpo-test-data/src/main/java/com/jporm/test/domain/section08/UserValue.java
@@ -35,7 +35,7 @@ public interface UserValue {
 	Optional<Long> id();
 
 	@Column(name = "AGE")
-	Long userAge();
+	Integer userAge();
 	String firstname();
 	String lastname();
